+++ Sources/KeyboardKit/Gestures/RepeatGestureTimer.swift
@@ -13,27 +13,49 @@ import Foundation
  keyboard button.
  */
 public class RepeatGestureTimer {
+
+    public init() {}
     
     deinit { stop() }
-    
-    public static let shared = RepeatGestureTimer()
-    
+
+
     private var timer: Timer?
     
     private var startDate: Date?
+
+    /**
+     The timer tick interval.
+     */
+    public var timeInterval: TimeInterval = 0.1
 }
 
 public extension RepeatGestureTimer {
-    
+
+    /**
+     This shared timer can be used if you only need to track
+     a single thing.
+     */
+    static let shared = RepeatGestureTimer()
+}
+
+public extension RepeatGestureTimer {
+
+    /**
+     The time for how long the timer has been active.
+     */
     var duration: TimeInterval? {
         guard let date = startDate else { return nil }
         return Date().timeIntervalSince(date)
     }
-    
+
+    /**
+     Whether or not the timer is active.
+     */
     var isActive: Bool { timer != nil }
-    
-    var timeInterval: TimeInterval { 0.1 }
-    
+
+    /**
+     Start the timer.
+     */
     func start(action: @escaping () -> Void) {
         if isActive { return }
         stop()
@@ -42,7 +64,10 @@ public extension RepeatGestureTimer {
             withTimeInterval: timeInterval,
             repeats: true) { _ in action() }
     }
-    
+
+    /**
+     Stop the timer.
+     */
     func stop() {
         timer?.invalidate()
         timer = nil